Extracted cell specifics into their own file, and introduced BooleanCell
